IRIS-3881 | RegisterPage unified
